hook up iconv.R, restrict parts to UTF-8 locales
git-svn-id: https://svn.r-project.org/R/trunk@81806 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ tests/iconv.R
@@ -23,11 +23,13 @@ stopifnot(iconv(z, "UTF-8", "ASCII", "Unicode") == "<U+0001F600>")
 stopifnot(iconv(z, "UTF-8", "ASCII", "c99") == "\\U0001f600")
 
 ## write out to compare with GNU libiconv's iconv on e.g. macOS
-if(startsWith(extSoftVersion()["iconv"], 'GNU libiconv')) {
+## The reading can only work in a UTF-8 locale
+if(startsWith(extSoftVersion()["iconv"], 'GNU libiconv') &&
+   l10n_info()[["UTF-8"]]) {
     writeLines(c(xx, z), "test.txt")
     zz <- system2("iconv", c("-f", "UTF-8", "-t", "c99", "test.txt"),
                   stdout = TRUE)
     unlink('test.txt')
     stopifnot(zz == iconv(c(xx, z), "UTF-8", "ASCII", "c99"))
     message('sub = "c99" agrees with GNU libiconv')
-}
+} else message('sub = "c99" agrees with GNU libiconv -- SKIPPED')
